WHAT IS YOUR PROBLEM GITHUB FOR WINDOWS????
When I say Commit this whole project -- I mean it. Not "OK, commit everything, except that folder -- even though I told you to include it". THIS is why this project hasn't been in Git since the beginning. I do not want to be wasting time babysitting Git and it's temper tantrums.