bluString: Partially implemented `Format` method
This converts the variant array into strings and pointers, ready for a
search and replace. This method will straight replace each successive
instance of a marker (probably "$") with the string version of the next
parameter in the array.

All this initial Variant & string-conversion work won't be the fastest,
but it'll certainly be faster than String concatenation, especially if
you reuse the same format string over and over.