Replaced "bluStringArray" with "bluArrayString"
This is more practical. One tends not to need to greatly modify strings
once already in an array, and the modifications could be done externally
by a bluString; storing lots of bluStrings is not more efficient than
simple VB Strings, and this new approach is more befitting what I need
to do in OZ80.

* bluArrayString now handles the conversion from Variant to String
instead of bluString