Some crazy re-architecting
Please excuse the crazy whilst I figure out how to architect this thing.
I'm looking at switching to a TLB for the WIN32 APIs, which would reduce
the amount of code and greatly speed up API calls.

A lot of work was done to bluString and there's some nice FileSystem
stuff going on, plus a couple of new functions to get the command line
and working directory easily.